add initial components for external user function
This change adds the mechanics to specify a URI for a file to be run as
part of the structured streaming pipeline. The function should take a
string and return a string, and be named `main`. The URI can be
specified by adding the environment variable `USER_FUNCTION_URI` to the
source-to-image build.